refactor: Http 프로토콜 삭제
