--- src/main/java/constant/HttpProtocol.java
@@ -1,4 +1,0 @@
-package constant;
-
-public class HttpProtocol {
-}
