+++ src/main/java/model/annotation/Api.java
@@ -1,6 +1,6 @@
 package model.annotation;
 
-import constant.HttpMethod;
+import model.enumeration.HttpMethod;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
