+++ src/main/java/model/enumeration/HttpMethod.java
@@ -1,4 +1,4 @@
-package constant;
+package model.enumeration;
 
 public enum HttpMethod {
     GET,
